Removed '#define i_private u.generic_ip' directive
The inode structure in the linux 2.4.18 kernel has a union field for file
system specific information. The inode structure in the linux 2.6/3.0 kernel
no longer has this union field. Instead, it seems like the 'void *i_private'
field in the new structure is there to serve the same purpose. Although, I was
unable to find any documentation to support this hunch.