Modified the 's_private' macro found in ux_fs.h
The super block structure in the linux 2.4.18 kernel has a union field for file
system specific information. The super block structure in the linux 2.6/3.0
kernel no longer has this union field. Instead, it seems like the
'void *s_fs_info' field in the new structure is there to serve the same
purpose. Although I was unable to find any documentation to support this hunch.